requirements.txt
streamlit==1.28.2
Pillow==8.3.2
numpy==1.21.2
tensorflow==2.5.0
